make metrics work with lists of dicts (#77)
With this PR, predicted or reference entries being lists/tuples
consisting of dictionaries get converted to (sets of) tuples so that
they work with set operations in the metrics.

This implements #69.

details:
- outsource method `_prepare_entry_as_set` (via new class
`MetricWithPrepareEntryAsSet`) from `ConfusionMatrix` / `F1Metric`
- **ignore `None` values** in multi-values (list **entries**) and in
compounds (dict **values**)
- add parameter `ignore_subfields` to metrics: An optional dict mapping
field names to lists of subfield names to ignore when converting dicts
to tuples (which are used for matching). This should be useful to
exclude e.g. `description` from compound fields (which should not be
evaluated since it is free text)