improve cluster startup (#54)
## goal
we want to be able to host an llm and run python scripts against it all
in one tidy command

## changes
- added `run_with_llm.sh` to host vllm on a compute node and run uv
against it on the same compute node
- if the anything crashes, the slurm job is stopped automatically.
manually stopping either the script or slurm job also stops the other.
- uv is run on the cluster node, which can take heavy compute tasks, but
may be more difficult to debug.
- added `run_with_llm_local_node_exec.sh` to host vllm on a compute node
and run uv against it on the login node
- if the script crashes, the slurm job needs to be stopped manually, and
vice versa
  - uv is run on the login node, which may not be desireable
- fixed flag in instructions on how to use vllm on the cluster

---------

Co-authored-by: Arne Binder <arne.binder@dfki.de>