fix warnings during inference (#93)
## goal 
Get rid of the following warning when doing inference with predict.py.
```
.../.venv/lib/python3.11/site-packages/pydantic/_internal/_generate_schema.py:2249: 
UnsupportedFieldAttributeWarning: 
The 'validate_default' attribute with value True was provided to the `Field()` function,
which has no effect in the context it was used.
'validate_default' is field-specific metadata, 
and can only be attached to a model field using `Annotated` metadata or by assignment.
This may have happened because an `Annotated` type alias using the `type` statement was used, 
or if the `Field()` function was attached to a single member of a union type.
  warnings.warn(
```
## minimal reproducible example
```sh
srun --partition=RTXA6000-SLT 
     --job-name=vllm_serve 
     --nodes=1 
     --ntasks=1      
     --cpus-per-task=6      
     --gpus-per-task=1      
     --mem-per-cpu=4G      
     --time=0-02:00:00      
     uvx --cache-dir /netscratch/$USER/cache/uv             
     vllm serve "openai/gpt-oss-20b"                        
     --download-dir=/ds/models/llms/cache                        
     --port=18000
export LLM_API_BASE="http://YOUR-SERVER.kl.dfki.de:18000/v1"
uv run -m kibad_llm.predict pdf_directory=./tests/fixtures/pdfs/
```
## changes
### upgrade python packages
```sh
uv lock --upgrade
```
the warning is related to our usage of Fields() in
`.src/kibad_llm/schema/types.py`. upgrading the python packages in our
uv.lock file fixes the warning.
### change pymupdf4llm markdown test
as we upgrade the packages, pymupdf4llm improves its performance in
converting pdf to markdown. for that i adjusted the testfile.<br>
the most noticeable improvement is its new ability to write co2emissions
instead of co 2 emissions when the 2 is in the to-the-power-of notation.