improve vllm documentation (#111)
Our documentation of how to use vLLM needed to be improved for stability
as well as usability. The new docs should be safer and easier to use.

details:
- Using `uv run -w vllm` can lead to dependecy hell, whereas a clean and
separate `uvx` is plenty
- Setting the default runtime of the slurm job to a full day has the
potential to be stuck in queues longer, as well as be forgotten and
block resources for unnecessarily long periods of time. Hencewhy, a
shorter default time of two hours is set.
- By calling `vllm@latest` it is clearer that this command will always
use the latest stable version of vLLM. This also can help people looking
for how to set a specific vLLM version.