improve models readme (#144)
The way we are using uv, slurm, and its related components like vllm and
huggingface have been changing. This PR aims to improve the instructions
on how to deal with our tooling on the pegasus cluster.

details:
- removed the instruction to create the cache dir for uv in the
prerequisites. uv takes care of that.
- added two env variables to the quick start to manage hf and vllm
caches.
- adapted the single package srun section to `uvx` instead of `uv run`
to reflect the previous change in the quickstart section. The original
change was done in #111, as `uvx` prevents many dependency hell issues
one may run into with `uv run`. Now our instructions are more
consistent.
- added vllm and hf to the symlinking section of full project srun for
completeness.
- added the `--vllm-version` flag info the the `run_with_llm.sh` docs.