extractor: make llm optional (#172)
With this PR, we do not fall back to `Settings.llm`, but skip the llm
call at all if no llm is provided and show a respective warning. The
motivation is to have a "dummy" prediction to easily see the full
prompt.

usage: set `extractor.llm=null` when calling `predict`

additional changes/notes:
- Implement `warn_once`: Log a warning message only once by caching the
function call.
 - The mentioned fallback was never used.