extractor: refactor message building (#179)
Especially, outsource `build_chat_messages` and `build_chat_message` and
add a `history` (messages) parameter. The latter is in preparation for
#180.

additional relevant changes:
- add parameters `schema_description_placeholder` and
`document_placeholder` and use them to check if the system or user
message should be formatted with either the system description or the
document text (and thus remove boolean parameter
`system_message_requires_schema_description`)
- make `system_message` or `user_message` optional (but require that at
least one is present). This will allow to omit one of them when
`history` is provided

This was successfully tested locally (slow tests) and in combination
with #180 (predict with `extractor=conditional_union_with_schema`).