extractor: prepare for vllm 0.12.0 (#189)
This implements #174.

details:
- use `"structured_outputs": {"json": schema}` instead of
`"guided_json": schema`
- parametrize `extra_body`: this contains additional parameters send to
the vllm backend
- create a new config subfolder `extractor/extra_body` with a single
`default.yaml` for now (same parameters as before)
- remove parameter `guided_decoding_backend` (with default
`lm-format-enforcer`)
- thus, `auto` will be used implicitly as guided decoding backend
(**breaking**)
- you can add
`+extractor.extra_body.guided_decoding_backend=lm-format-enforcer` to
restore the previous behavior. But this won't work in the future (as I
understand it, with `vllm>=0.12.0` vllm needs to be started explicitly
with a certain backend. but this needs to be verified)