extractor: prepare for different backend types (#199)
This is in preparation for #200 and #201. Especially, guided decoding
needs different ways to pass the schema for different backend types (and
depending on if it is self-hosted via vLLM or from official OpenAI), so
we explicitly define it per type (and introduce the new type
`kibad_llm.models.OpenAILikeVllm` to indicate that our gpt-oss-20b is
hosted via vLLM).

details:
 - outsource `_call_llm_chat_with_guided_decoding`
- rename `extra_body` to `request_parameters` (since `extra_body` is
vllm specific) (**breaking!**)
- add `kibad_llm.models.OpenAILikeVllm` wrapper class and use it instead
of `OpenAILike` to indicate vllm based hosting
- IMPORTANT: we raise an error, if an llm (interface) of unknown type is
used; for now only `OpenAILikeVllm` is known (**breaking!**)