refactor llm interfaces (#205)
This is a follow-up of #199. Especially, this moves the logic how to
call the individual llm interfaces to the llm interface classes. This is
tested with #202.

details:
 - move `models` to `llms` package
- add `llms.base.LLM` interface with abstract
`call_llm_chat_with_guided_decoding` and some helper methods:
   - `get_response_content_from_chat_response`
   - `get_raw_message_from_chat_response`
   - `get_reasoning_from_chat_response`
- remove `_call_llm_chat_with_guided_decoding` from `extractor.base`,
but call `llm.call_llm_chat_with_guided_decoding`
- add `llms.base.SimpleChatMessage` and use it instead of `ChatMessage`
from Llama Index to simplify the interface
 - move exceptions to `llms.base.LLM`
 - move `warn_once` to `utils.log`
 - implement `OpenAILikeVllm.call_llm_chat_with_guided_decoding`