implement `ConditionalUnionExtractor` and add configs (#180)
This implements #70 by simply extending the [`union`
extractor](https://github.com/DFKI-NLP/kibad-llm/blob/main/configs/extractor/union_with_schema.yaml):
the history of previous messages is collected and passed to the next LLM
queries. However, the (initial) system message is passed only once, it
is expected that it contains the input document. Note that this was in
the user message in our previous setups!