Load dataset with overrides (#221)
... and integrate them into the output for evaluate commands. This
implements #218.

details:
- implement `dataset.prediction.load_with_metadata`: Load a dataset from
a prediction log directory, extracting metadata such as Hydra overrides
and attach it to the dataset. If `log` is not provided, load directly
from `file` (backward compatibility).
- we add a complete run log dir for testing (in fact, just
`job_return_value.json` and `.hydra/overrides.yaml` would be required,
but I kept the other files to have a consistent setup if we ever want to
use more data than the `overrides`)
- use it per default in `extraction_result.yaml` config (which is the
default for all eval datasets)
- propagate `metadata` from predictions dataset to combined dataset (in
`merge_references_into_predictions`)
 - add the `dataset.metadata` to job result (in `evaluate.py`)
- adjust `SaveJobReturnValueCallback` to extend (or replace) the
original job overrides with the overrides from the job result
- add parameter `multirun_replace_existing_overrides` (default: True):
If True, replace existing overrides in the job return-value with the
overrides from the job return object if available. If False, the
overrides from the job return-value are only appended if no overrides
are available in the job return object.