Add prediction `job_return_value` to evaluation multirun result (#228)
This is a follow-up of #221. With this PR, all content of the
*prediction* result (the metadata: `output_file`, `time_extraction`,
`time_pdf_conversion`) will be saved in the *evaluate* result when
called as multirun. 

Note that this is not added to the single run output since it may
destroy its format, especially, for the markdown table. This is not a
problem for the multirun markdown table since it is flattened.